fix bugs

+++ Assets/3. Scripts/UI/Panels/PetsPanel.cs
@@ -39,11 +39,17 @@ namespace _3._Scripts.UI.Panels
             booster.gameObject.SetActive(false);
             remove.onClick.AddListener(Remove);
             selectBest.onClick.AddListener(SelectBest);
-            rewardButton.onClick.AddListener(() => GBGames.ShowRewarded(GetRandomPet));
+            rewardButton.onClick.AddListener(GetRandomPet);
         }
 
         public void ShowOffer()
         {
+            if (GBGames.saves.petsSave.MaxUnlocked(25))
+            {
+                NotificationPanel.Instance.ShowNotification("max_pet_unlocked");
+                return;
+            }
+            
             var panel = UIManager.Instance.GetPanel<OfferPanel>();
             var pets = Configuration.Instance.AllPets.ToList();
             var data = pets[Random.Range(0, pets.Count)];
@@ -63,18 +69,27 @@ namespace _3._Scripts.UI.Panels
 
         private void GetRandomPet()
         {
-            var pets = Configuration.Instance.AllPets.ToList();
-            var data = pets[Random.Range(0, pets.Count)];
-            var maxBooster = GBGames.saves.petsSave.GetMaxBooster();
-            var currentBooster = Random.Range(maxBooster, maxBooster + 5);
-
-            GBGames.saves.petsSave.Unlock(data, currentBooster);
-            PetUnlocker.SelectBest();
+            if (GBGames.saves.petsSave.MaxUnlocked(25))
+            {
+                NotificationPanel.Instance.ShowNotification("max_pet_unlocked");
+                return;
+            }
             
-            DeleteSlots();
-            InitializeSlots();
-            UpdateCount();
-            EquipBest();
+            GBGames.ShowRewarded(() =>
+            {
+                var pets = Configuration.Instance.AllPets.ToList();
+                var data = pets[Random.Range(0, pets.Count)];
+                var maxBooster = GBGames.saves.petsSave.GetMaxBooster();
+                var currentBooster = Random.Range(maxBooster, maxBooster + 5);
+
+                GBGames.saves.petsSave.Unlock(data, currentBooster);
+                PetUnlocker.SelectBest();
+            
+                DeleteSlots();
+                InitializeSlots();
+                UpdateCount();
+                EquipBest(); 
+            });
         }
 
         protected override void OnOpen()
